+++ backend/models/user.model.js
@@ -5,6 +5,9 @@ const userSchema = new mongoose.Schema({
   surname: { type: String, required: true },
   email: { type: String, required: true, unique: true },
   password: { type: String, required: true },
+  address: { type: String, required: true },
+  phone: { type: String, required: true },
+  isAdmin: { type: Boolean, required: true, default: false },
 })
 
 export default mongoose.model("User", userSchema)