+++ frontend/src/pages/auth/Login.jsx
@@ -1,35 +1,41 @@
+import { useState } from "react"
 import { useTranslation } from "react-i18next"
-import { Link, useParams } from "react-router-dom"
-import {
-  Container,
-  Row,
-  Col,
-  Card,
-  CardBody,
-  Form,
-  Button,
-} from "react-bootstrap"
+import { Link, useParams, useNavigate } from "react-router-dom"
+import { Container, Row, Col, Card, Form, Button } from "react-bootstrap"
+import { handleLogin } from "@/utils/users"
 
 const Login = () => {
   const { t } = useTranslation()
   const { lang } = useParams()
+  const navigate = useNavigate()
+
+  const [email, setEmail] = useState("")
+  const [password, setPassword] = useState("")
+  const [error, setError] = useState("")
+
+  const handleLoginClick = () => {
+    handleLogin(email, password, setError, navigate, lang)
+  }
 
   return (
-    <Container fluid>
-      <Row className="d-flex justify-content-center align-items-center">
-        <Col md={6} lg={4} xl={5}>
+    <Container>
+      <Row>
+        <Col>
           <Card
             className="bg-white mx-auto"
             style={{ borderRadius: "1rem", maxWidth: "500px" }}
           >
-            <CardBody className="p-5 w-100 d-flex flex-column">
-              <h2 className="fw-bold mb-5 text-center">{t("login.login")}</h2>
+            <Card.Body className="p-5 w-100 d-flex flex-column">
+              <h3 className="mb-5 text-center">{t("login.login")}</h3>
+              {error && <p className="text-danger text-center">{error}</p>}
               <Form>
                 <Form.Group className="mb-4" controlId="formEmail">
                   <Form.Control
                     type="email"
                     placeholder={t("login.placeholder.email")}
                     size="lg"
+                    value={email}
+                    onChange={(e) => setEmail(e.target.value)}
                   />
                 </Form.Group>
 
@@ -38,27 +44,19 @@ const Login = () => {
                     type="password"
                     placeholder={t("login.placeholder.password")}
                     size="lg"
+                    value={password}
+                    onChange={(e) => setPassword(e.target.value)}
                   />
-                </Form.Group>
-
-                <Form.Group className="mb-4" controlId="formCheckbox">
-                  <div className="d-flex justify-content-center">
-                    <Form.Check
-                      type="checkbox"
-                      label={t("login.remember_pw")}
-                    />
-                  </div>
                 </Form.Group>
 
                 <Button
                   variant="primary"
                   size="lg"
                   className="w-100 mb-4"
-                  onClick={() => alert("In development")}
+                  onClick={handleLoginClick}
                 >
                   {t("login.login")}
                 </Button>
-                <br />
                 <p className="text-center">{t("login.or")}</p>
                 <Link to={`/${lang}/register`}>
                   <Button variant="secondary" size="lg" className="w-100">
@@ -66,7 +64,7 @@ const Login = () => {
                   </Button>
                 </Link>
               </Form>
-            </CardBody>
+            </Card.Body>
           </Card>
         </Col>
       </Row>